+++ src/sites/github.ts
@@ -20,7 +20,7 @@ const github = new Site({
       `,
       manipulateDom: async ({ siteAction }) => {
         const container = await waitForElement('#dashboard')
-        const quote = siteAction.createQuoteElement(container)
+        const quote = siteAction.createQuoteWidget(container)
         if (!quote)
           return
         container.after(quote)
