+++ src/sites/tiktok.ts
@@ -22,7 +22,7 @@ const tiktok = new Site({
         const container = await waitForElement('[data-e2e="recommend-list-item-container"]')
         mute(container)
         setTimeout(() => {
-          const quote = siteAction.createQuoteElement(container)
+          const quote = siteAction.createQuoteWidget(container)
           if (!quote)
             return
           container.before(quote)
