+++ src/sites/twitter.ts
@@ -21,7 +21,7 @@ const twitter = new Site({
       manipulateDom: async ({ siteAction }) => {
         const container = await waitForElement('[data-testid="primaryColumn"] > div:last-child > div:nth-child(4)')
         mute(container)
-        const quote = siteAction.createQuoteElement(container)
+        const quote = siteAction.createQuoteWidget(container)
         if (!quote)
           return
         quote.style.padding = '40px'
@@ -39,7 +39,7 @@ const twitter = new Site({
       `,
       manipulateDom: async ({ siteAction }) => {
         const container = await waitForElement('[data-testid="sidebarColumn"] section')
-        const quote = siteAction.createQuoteElement(container)
+        const quote = siteAction.createQuoteWidget(container)
         if (!quote)
           return
         quote.style.padding = '20px'
@@ -57,7 +57,7 @@ const twitter = new Site({
       `,
       manipulateDom: async ({ siteAction }) => {
         const container = await waitForElement('[data-testid="sidebarColumn"] aside')
-        const quote = siteAction.createQuoteElement(container)
+        const quote = siteAction.createQuoteWidget(container)
         if (!quote)
           return
         quote.style.padding = '20px'
